requirements.txt
# RAG Chatbot Dependencies
# Core AI and Vector Database
cohere>=4.32
chromadb>=0.4.22
langchain>=0.1.5
langchain-cohere>=0.1.5
langchain-community>=0.0.19
langchain-text-splitters>=0.0.1

# Hugging Face Integration
transformers>=4.21.0
torch>=1.13.0
huggingface-hub>=0.14.0
langchain-huggingface>=0.0.1

# UI Framework
streamlit>=1.31.0

# Text Processing
tiktoken>=0.5.2

# Utilities
python-dotenv>=1.0.0

# Additional dependencies for document processing
unstructured>=0.10.0
pypdf>=3.0.0

# Additional requested dependencies
openai
faiss-cpu
requests